+++ libs/act/test/builder.spec.ts
@@ -5,9 +5,9 @@ const A1 = state("A1", z.object({}))
   .init(() => ({}))
   .emits({ Event1: z.object({}), Event2: z.object({}) })
   .patch({ Event1: () => ({}), Event2: () => ({}) })
-  .on("Act1", z.object({}))
+  .on({ Act1: z.object({}) })
   .emit(() => ["Event1", {}])
-  .on("Act2", z.object({}))
+  .on({ Act2: z.object({}) })
   .emit(() => ["Event2", {}])
   .build();
 
@@ -15,7 +15,7 @@ const A2 = state("A2", ZodEmpty)
   .init(() => ({}))
   .emits({ Event22: ZodEmpty })
   .patch({ Event22: () => ({}) })
-  .on("Act1", ZodEmpty)
+  .on({ Act1: ZodEmpty })
   .emit(() => ["Event22", {}])
   .build();
 
@@ -23,7 +23,7 @@ const A3 = state("A3", ZodEmpty)
   .init(() => ({}))
   .emits({ Event1: ZodEmpty, Event2: ZodEmpty })
   .patch({ Event1: () => ({}), Event2: () => ({}) })
-  .on("Act3", ZodEmpty)
+  .on({ Act3: ZodEmpty })
   .emit(() => ["Event1", {}])
   .build();
 
@@ -134,7 +134,7 @@ describe("Builder", () => {
       .init(() => ({}))
       .emits({ E: z.object({}) })
       .patch({ E: () => ({}) })
-      .on("A", z.object({}))
+      .on({ A: z.object({}) })
       .emit(() => ["E", {}])
       .build();
 
@@ -157,7 +157,7 @@ describe("Builder", () => {
       .init(() => ({}))
       .emits({ E: z.object({}) })
       .patch({ E: () => ({}) })
-      .on("A", z.object({}))
+      .on({ A: z.object({}) })
       .emit(() => ["E", {}])
       .build();